# Implement a faster version of `FromNormalizedStr` for `Name`s