@jv-garcia 2025-02-25
@cdisselkoen Not sure if you could take a look at this, I can't actively work on it, but it'd be great if you folks could at least have a conversation on whether the approach is correct, whether you think it's worth it, and what else we'd be missing (proptesting for the implementation perhaps?).