@shaobo-he-aws 2025-02-26
IMO, if we want performance improvement, why not do the extreme and not use regex? It appears to me that the parser can be implemented in a while loop.